experiments/013_text_summary2/requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.14.0
peft>=0.10.0
accelerate>=0.27.0
deepspeed>=0.13.0

# Inference
vllm>=0.6.0

# Logging & tracking
wandb>=0.16.0

# Evaluation metrics
rouge-score>=0.1.2
bert-score>=0.3.13

# NLI faithfulness check
sentence-transformers>=2.2.0

# LLM-as-judge
anthropic>=0.39.0

# PDF extraction
pymupdf>=1.24.0

# Config
pyyaml>=6.0

# Flash Attention 2 (requires CUDA 11.8+)
flash-attn>=2.5.0

# Build tools
ninja
packaging
